+++ game.js
@@ -7,51 +7,29 @@ var currentState = undefined;
 var currentMasterVolume = 1;
 var muted = false;
 
-var touch = {
-    holdX: undefined,
-    holdY: undefined,
-    currentX: undefined,
-    currentY: undefined,
-    angle: function()
-    {
-        if(isTouchDragged())
-        {
-            return Math.atan2(this.currentY - this.holdY, this.currentX - this.holdX)
-        }
-
-        return undefined
-    }
-}
-
-const controls = {
-    up: false,
-    down: false,
-    right: false,
-    left: false
-}
+var touch;
+var controls;
 
 const camera = {
     x: 0,
     y: 0
 }
 
-function preload() 
-{
+function preload() {
     loadResources()
 }
 
-function setup() 
-{
+function setup() {
     createCanvas(windowWidth, windowHeight);
     frameRate(60);
-
+    touch = new Touch();
+    controls = new Controls();
     //registeredStates.push(new MainMenuState());
     registeredStates.push(new InGameState());
     currentState = registeredStates[0];
 }
 
-function draw() 
-{
+function draw() {
     background(0);
     logic()
     render()
@@ -59,58 +37,45 @@ function draw()
 }
 
 
-function logic() 
-{
-    if(currentState)
-    {
+function logic() {
+    if (currentState) {
         currentState.updateState();
     }
 }
 
-function render() 
-{
-    if(currentState)
-    {
+function render() {
+    if (currentState) {
         currentState.renderState()
     }
 }
 
-function renderUI() 
-{
-    if(currentState)
-    {
+function renderUI() {
+    if (currentState) {
         currentState.renderUIState();
     }
 }
 
-function keyPressed() 
-{
-    if(currentState)
-    {
+function keyPressed() {
+    if (currentState) {
         currentState.onKeyPressed();
     }
 
     return false;
 }
 
-function keyReleased() 
-{
-    if(currentState)
-    {
+function keyReleased() {
+    if (currentState) {
         currentState.onKeyReleased();
     }
 
-    if(key === 'm')
-    {
+    if (key === 'm') {
         muted = !muted
 
-        if(muted)
-        {
+        if (muted) {
             masterVolume(0)
         }
 
-        else
-        {
+        else {
             masterVolume(currentMasterVolume)
         }
     }